requirements.txt
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.16.0
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.2.0
coverage==7.3.2
cryptography==42.0.2
cycler==0.12.1
debugpy==1.6.5
decorator==5.1.1
dill==0.3.8
emoji==2.2.0
entrypoints==0.4
executing==1.2.0
Flask==2.2.3
fonttools==4.48.1
gunicorn==21.2.0
idna==3.4
ipykernel==6.20.2
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.9
jupyter_core==5.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
MouseInfo==0.1.3
mpmath==1.3.0
Naked==0.1.32
nest-asyncio==1.5.6
numpy==1.24.1
packaging==23.0
parso==0.8.3
pbr==6.0.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
ply==3.11
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.20.0
pygame==2.3.0
PyGetWindow==0.0.9
Pygments==2.14.0
pylatexenc==2.10
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyRect==0.2.0
PyScreeze==0.1.28
PySide6==6.4.2
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
pyspnego==0.10.2
python-dateutil==2.8.2
pytweening==1.0.4
pywhatkit==5.4
pywin32==305
PyYAML==6.0.1
pyzmq==25.0.0
qiskit==0.46.0
qiskit-aer==0.13.3
qiskit-ibm-provider==0.9.0
qiskit-terra==0.46.0
requests==2.28.2
requests-ntlm==1.2.0
rustworkx==0.14.0
scapy==2.5.0
scipy==1.12.0
shellescape==3.8.1
shiboken6==6.4.2
six==1.16.0
soupsieve==2.3.2.post1
sspilib==0.1.0
stack-data==0.6.2
stevedore==5.1.0
sympy==1.12
tornado==6.2
traitlets==5.8.1
typing_extensions==4.9.0
urllib3==1.26.14
wcwidth==0.2.6
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.2.3
wikipedia==1.4.0
